# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Наша CRM",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/activity/category/{category_id}": {
            "get": {
                "tags": [
                    "Занятия"
                ],
                "summary": "Получить категорию по id",
                "description": "Вернет категории по id",
                "operationId": "2e63eb3d994b87b60b9b79c1941fd941",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID категории",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "test"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/activity/category": {
            "get": {
                "tags": [
                    "Занятия"
                ],
                "summary": "Получить список категорию занятий",
                "description": "Вернет список всех категорий занятий текущего пользователя",
                "operationId": "cb8ea88ab43ce694462c61f0a21f2bbe",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "Fanger"
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "Fanger"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/activity/category/add": {
            "post": {
                "tags": [
                    "Занятия"
                ],
                "summary": "Добавить категорию занятия",
                "description": "Нужно передать поля name",
                "operationId": "23706b9be4bf1641c1ef1abcf4c7f2aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/activity/category/update/{category_id}": {
            "put": {
                "tags": [
                    "Занятия"
                ],
                "summary": "Обновить категорию занятия",
                "description": "Нужно передать поля name",
                "operationId": "04e9292272941d41947e381a96d26bbd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}